Fixed comments after code review
