+++ test/list_test.dart
@@ -20,16 +20,15 @@
 
 import 'dart:io';
 import 'package:test/test.dart' hide test, throws;
-import 'test_base.dart';
+import 'test.dart';
 import '../lib/realm.dart';
-import 'test_model.dart';
 
 Future<void> main([List<String>? args]) async {
   parseTestNameFromArguments(args);
 
   print("Current PID $pid");
 
-  setupTests(Configuration.filesPath, (path) => {Configuration.defaultPath = path});
+  setupTests();
 
   test('Lists create object with a list property', () {
     var config = Configuration([Team.schema, Person.schema]);
@@ -265,7 +264,7 @@ Future<void> main([List<String>? args]) async {
     realm.close();
   });
 
-  test('Access deleted object collection', () {
+  test('Read list property of a deleted object', () {
     var config = Configuration([Team.schema, Person.schema]);
     var realm = Realm(config);
 
@@ -277,7 +276,7 @@ Future<void> main([List<String>? args]) async {
     realm.close();
   });
 
-  test('Delete collection of deleted parent', () {
+  test('Delete a list of objects through a deleted parent', () {
     var config = Configuration([Team.schema, Person.schema]);
     var realm = Realm(config);
 
@@ -287,45 +286,9 @@ Future<void> main([List<String>? args]) async {
     realm.write(() => realm.delete(team));
     expect(() => realm.write(() => realm.deleteMany(players)), throws<RealmException>("Access to invalidated Collection object"));
     realm.close();
-  });
+  });  
 
-  test('Realm deleteMany from list after realm is closed', () {
-    var config = Configuration([Team.schema, Person.schema]);
-    var realm = Realm(config);
-
-    //Create a team
-    final team = Team("Ferrari");
-    realm.write(() => realm.add(team));
-
-    //Add players to the team
-    final newPlayers = [
-      Person("Michael Schumacher"),
-      Person("Sebastian Vettel"),
-      Person("Kimi Räikkönen"),
-    ];
-    realm.write(() => team.players.addAll(newPlayers));
-
-    //Ensure team exists in realm
-    var teams = realm.all<Team>();
-    expect(teams.length, 1);
-
-    //Try to delete team players while realm is closed
-    final players = teams[0].players;
-    realm.close();
-    expect(
-        () => realm.write(() {
-              realm.deleteMany(players);
-            }),
-        throws<RealmException>());
-
-    //Ensure all persons still exists in realm
-    realm = Realm(config);
-    final allPersons = realm.all<Person>();
-    expect(allPersons.length, 3);
-    realm.close();
-  });
-
-  test('Get list length after deleting parent objects', () {
+  test('Get length of list property on a deleted object', () {
     var config = Configuration([Team.schema, Person.schema]);
     var realm = Realm(config);
 
@@ -361,96 +324,7 @@ Future<void> main([List<String>? args]) async {
     expect(players.isValid, false);
   });
 
-  test('Realm deleteMany from list referenced by two objects', () {
-    var config = Configuration([Team.schema, Person.schema]);
-    var realm = Realm(config);
-
-    //Create two teams
-    final teamOne = Team("Ferrari");
-    final teamTwo = Team("Maserati");
-    realm.write(() {
-      realm.add(teamOne);
-      realm.add(teamTwo);
-    });
-
-    //Create common players list for both teams
-    final newPlayers = [
-      Person("Michael Schumacher"),
-      Person("Sebastian Vettel"),
-      Person("Kimi Räikkönen"),
-    ];
-    realm.write(() {
-      teamOne.players.addAll(newPlayers);
-      teamTwo.players.addAll(newPlayers);
-    });
-
-    //Ensule teams exist in realm
-    var teams = realm.all<Team>();
-    expect(teams.length, 2);
-
-    //Delete all players in a team from realm
-    realm.write(() => realm.deleteMany(teams[0].players));
-
-    //Ensure all players are deleted from collection
-    expect(teams[0].players.length, 0);
-
-    //Reload all persons from realm and ensure they are deleted
-    final allPersons = realm.all<Person>();
-    expect(allPersons.length, 0);
-    realm.close();
-  });
-
-  test('RealmObject add with list properties', () {
-    var config = Configuration([Team.schema, Person.schema]);
-    var realm = Realm(config);
-
-    final team = Team("Ferrari")
-      ..players.addAll([Person("Michael"), Person("Kimi")])
-      ..scores.addAll([1, 2, 3]);
-
-    realm.write(() => realm.add(team));
-
-    final teams = realm.all<Team>();
-    expect(teams.length, 1);
-    expect(teams[0].name, "Ferrari");
-    expect(teams[0].players, isNotNull);
-    expect(teams[0].players.length, 2);
-    expect(teams[0].players[0].name, "Michael");
-    expect(teams[0].players[1].name, "Kimi");
-
-    expect(teams[0].scores.length, 3);
-    expect(teams[0].scores[0], 1);
-    expect(teams[0].scores[1], 2);
-    expect(teams[0].scores[2], 3);
-    realm.close();
-  });
-
-  test('Realm deleteMany from iterable', () {
-    var config = Configuration([Team.schema, Person.schema]);
-    var realm = Realm(config);
-
-    //Create two teams
-    final teamOne = Team("Team one");
-    final teamTwo = Team("Team two");
-    final teamThree = Team("Team three");
-    realm.write(() {
-      realm.add(teamOne);
-      realm.add(teamTwo);
-      realm.add(teamThree);
-    });
-
-    //Ensure the teams exist in realm
-    var teams = realm.all<Team>();
-    expect(teams.length, 3);
-
-    //Delete teams one and three from realm
-    realm.write(() => realm.deleteMany([teamOne, teamThree]));
-
-    //Ensure both teams are deleted and only teamTwo has left
-    expect(teams.length, 1);
-    expect(teams[0].name, teamTwo.name);
-    realm.close();
-  });
+  
 
   test('Realm deleteMany from realm list', () {
     var config = Configuration([Team.schema, Person.schema]);
@@ -481,6 +355,44 @@ Future<void> main([List<String>? args]) async {
     //Reload all persons from realm and ensure they are deleted
     final allPersons = realm.all<Person>();
     expect(allPersons.length, 0);
+    realm.close();
+  });
+
+  test('List notifications', () async {
+    var config = Configuration([Team.schema, Person.schema]);
+    var realm = Realm(config);
+
+    final team = Team('t1', players: [Person("p1")]);
+    realm.write(() => realm.add(team));
+
+    var firstCall = true;
+    final subscription = (team.players as RealmList<Person>).changes.listen((changes) {
+      if (firstCall) {
+        firstCall = false;
+        expect(changes.inserted.isEmpty, true);
+        expect(changes.modified.isEmpty, true);
+        expect(changes.deleted.isEmpty, true);
+        expect(changes.newModified.isEmpty, true);
+        expect(changes.moved.isEmpty, true);
+      } else {
+        expect(changes.inserted, [1]); //new object at index 1
+        expect(changes.modified, [0]); //object at index 0 changed
+        expect(changes.deleted.isEmpty, true);
+        expect(changes.newModified, [0]);
+        expect(changes.moved.isEmpty, true);
+      }
+    });
+
+    await Future<void>.delayed(Duration(milliseconds: 10));
+    realm.write(() {
+      team.players.add(Person("p2"));
+      team.players.first.name = "p3";
+    });
+
+    await Future<void>.delayed(Duration(milliseconds: 10));
+    subscription.cancel();
+
+    await Future<void>.delayed(Duration(milliseconds: 10));
     realm.close();
   });
 }
