+++ test/test.dart
@@ -24,6 +24,9 @@ import 'package:meta/meta.dart';
 import 'package:path/path.dart' as _path;
 import 'package:test/test.dart';
 import 'package:test/test.dart' as testing;
+import '../lib/realm.dart';
+
+part 'test.g.dart';
 
 String? testName;
 
@@ -58,13 +61,13 @@ void parseTestNameFromArguments(List<String>? arguments) {
   }
 }
 
-void setupTests(String filePath, void Function(String path) setDefaultPath) {
+void setupTests() {
   setUp(() {
     String path = "${generateRandomString(10)}.realm";
     if (Platform.isAndroid || Platform.isIOS) {
-      path = _path.join(filePath, path);
+      path = _path.join(Configuration.filesPath, path);
     }
-    setDefaultPath(path);
+    Configuration.defaultPath = path;
 
     addTearDown(() async {
       var file = File(path);
@@ -104,4 +107,51 @@ Future<void> tryDeleteFile(FileSystemEntity fileEntity, {bool recursive = false}
       await Future<void>.delayed(Duration(milliseconds: 50));
     }
   }
+}
+
+@RealmModel()
+class _Car {
+  @PrimaryKey()
+  late String make;
+}
+
+@RealmModel()
+class _Person {
+  late String name;
+}
+
+@RealmModel()
+class _Dog {
+  @PrimaryKey()
+  late String name;
+
+  late int? age;
+
+  _Person? owner;
+}
+
+@RealmModel()
+class _Team {
+  late String name;
+  late List<_Person> players;
+  late List<int> scores;
+}
+
+@RealmModel()
+class _Student {
+  @PrimaryKey()
+  late int number;
+  late String? name;
+  late int? yearOfBirth;
+  late _School? school;
+}
+
+@RealmModel()
+class _School {
+  @PrimaryKey()
+  late String name;
+  late String? city;
+  List<_Student> students = [];
+  late _School? branchOfSchool;
+  late List<_School> branches;
 }
